requirements.txt
boto3==1.36.7
Flask==3.1.0
python-dotenv==0.21.1
jira==3.8.0
pydantic==2.10.6
prometheus-flask-exporter==0.22.0
gunicorn==23.0.0
setuptools==75.8.0
pytest==8.3.4
moto==5.0.28